RAG-ClinicalBERT(nadella)/requirements.txt
# Clinical Insight Bot — Python Dependencies
# Tested with Python 3.10–3.13

# Constrain numpy for compatibility
numpy<2

# Database and vector search
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Data processing
pandas>=2.0.3
scikit-learn>=1.3.0

# ClinicalBERT and ML
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu

# Web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
flask>=2.3.3
flask-cors>=4.0.0
flask-login>=0.6.3

# Text processing
tiktoken>=0.5.1
nltk>=3.8.1
textblob>=0.17.1

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0

# OpenAI (answer generation)
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.1
